% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTranscript.R
\name{checkTranscript}
\alias{checkTranscript}
\title{checks the transcript and validates the name convention, splits the name and checks if it is Species_Type.DataBase.Version.optional.fa.gz}
\usage{
checkTranscript()
}
\value{
returns a validation 0,1 where 1 is failure
}
\description{
checks the transcript and validates the name convention, splits the name and checks if it is Species_Type.DataBase.Version.optional.fa.gz
}

