% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAppResultOutputDirectory.R
\name{createAppResultOutputDirectory}
\alias{createAppResultOutputDirectory}
\title{create an ouptut Directory with correct paths for uploading to basespace this is for single node, more for analysis phase II.  also creates a scratch directory hierarchy.}
\usage{
createAppResultOutputDirectory(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{list of input ids created by formbuilder used to parse json file}

\item{fileJSON}{file json read from appSession}
}
\value{
outputPath   a path to output directory
}
\description{
create an ouptut Directory with correct paths for uploading to basespace this is for single node, more for analysis phase II.  also creates a scratch directory hierarchy.
}

