% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDesignMatrixFromAppSession.R
\name{createDesignMatrixFromAppSession}
\alias{createDesignMatrixFromAppSession}
\title{create the Design matrix from comparison and control inputs on basespace}
\usage{
createDesignMatrixFromAppSession(appSession)
}
\arguments{
\item{selectNames}{vector of input ids from baseSpace input form}
}
\value{
design      matrix of a one group comparison default
}
\description{
create the Design matrix from comparison and control inputs on basespace
}

