% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutputDirectory.R
\name{createOutputDirectory}
\alias{createOutputDirectory}
\title{create output directory for single node computations}
\usage{
createOutputDirectory(selectNames)
}
\arguments{
\item{selectNames}{vector of input id's from basespace input form}
}
\value{
outputPath    string of basespace output directory for uploading
}
\description{
create output directory for single node computations
}

