% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erccAnalysis.R
\name{erccAnalysis}
\alias{erccAnalysis}
\title{QC plots of ERCC spike-in controls this is targeted for PHase II inputing an appSession list}
\usage{
erccAnalysis(kexp, appSession)
}
\arguments{
\item{kexp}{something that behaves like a KallistoExperiment 
@param appSession     a list of analysis workflow items, including comparison names and control names}
}
\value{
returns plots and images of ercc analysis
}
\description{
QC plots of ERCC spike-in controls this is targeted for PHase II inputing an appSession list
}

