% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbiasPseudobam.R
\name{findbiasPseudobam}
\alias{findbiasPseudobam}
\title{find the bias or pseudobam input from the basespace interface}
\usage{
findbiasPseudobam(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{vector of input ids from basespace form}

\item{fileJSON}{json file from basespace}
}
\value{
biasPBList    list of bias or pseudobam flags used for quant
}
\description{
find the bias or pseudobam input from the basespace interface
}

