% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findkmerSize.R
\name{findkmerSize}
\alias{findkmerSize}
\title{find the kmer size input from basespace user input}
\usage{
findkmerSize(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{vector of input ids used to parse the json file}

\item{fileJSON}{json file from basespace}
}
\value{
kmerSize      numeric number of the kmer user to index reference
}
\description{
find the kmer size input from basespace user input
}

