% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSingleNodeProjectId.R
\name{findSingleNodeProjectId}
\alias{findSingleNodeProjectId}
\title{finds the project ID from the input basespace, required for uploading back, this is for single node}
\usage{
findSingleNodeProjectId(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{vector of strings of input ids from basespace}

\item{fileJSON}{json file used to parse}
}
\value{
projectID      string numeric of project id used for uploading back to baseSpace
}
\description{
finds the project ID from the input basespace, required for uploading back, this is for single node
}

