% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTranscriptInput.R
\name{findTranscriptInput}
\alias{findTranscriptInput}
\title{find the transcript input for single node creates hash table from storing into environment internally}
\usage{
findTranscriptInput(selectNames)
}
\arguments{
\item{selectNames}{vector of input ids used to parse the json file}
}
\value{
transcriptome  vector of input transcripts
}
\description{
find the transcript input for single node creates hash table from storing into environment internally
}

