% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAnalysisWorkFlow.R
\name{processAnalysisWorkflow}
\alias{processAnalysisWorkflow}
\title{arkas analysis is Phase II for basespace cloud compting algorithm it uses a single node so arkasJSONparse.R will parse the single node JSON and prepare for the workflow. the appresults are downloaded into the /data/input/appresults directory, in which there will be 1 directory for each appResults selected by the user (fasta files, comparison, and control results). The custom fasta form id : app-result-id,  the control form id: control-app-id,  the comparison form id: comparison-app-id.  these id names should not change ,so the JSON parsing won't be affected by adding new fields to form-builder.}
\usage{
processAnalysisWorkflow()
}
\value{
returns a list of appSession with the correct elements for running arkas workflow in entirety
}
\description{
arkas analysis is Phase II for basespace cloud compting algorithm it uses a single node so arkasJSONparse.R will parse the single node JSON and prepare for the workflow. the appresults are downloaded into the /data/input/appresults directory, in which there will be 1 directory for each appResults selected by the user (fasta files, comparison, and control results). The custom fasta form id : app-result-id,  the control form id: control-app-id,  the comparison form id: comparison-app-id.  these id names should not change ,so the JSON parsing won't be affected by adding new fields to form-builder.
}

